SVS-03: fail closed on duplicate sources and self-verify every emitted artifact
A duplicated source id is refused on both sides (the build's
DuplicateSourceIdentity, the verifier's DuplicateSourceMeta) instead of
letting meta_for silently bind the first entry. verify_artifact_with_sources
is the combined entry point — artifact verification plus source digests in
one call — and build_package now runs it against its own output under a
trust root derived from its signing key, so a build can never emit an
artifact the full verifier would reject. That self-check exposed real gaps
it now closes: a source whose every record is voided, clipped, or loses
deterministic resolution gets an explicit NoContribution disposition, so a
consumed source is never a phantom summary; and a shared source-record
reference now fails the build itself. Source verification moved to
verify/sources with its tests in verify/tests/source_binding.
