+++ crates/pilotage-svs-build/src/lib.rs
@@ -57,4 +57,7 @@ pub use source::{
     Aerodrome, LicenseCode, Obstacle, ObstacleKind, Runway, SourceDataset, SourceId, SourceMeta,
     SourceRecordRef, TerrainGrid,
 };
-pub use verify::{DecodedReports, decode_package_reports, verify_artifact, verify_source_digests};
+pub use verify::{
+    DecodedReports, decode_package_reports, verify_artifact, verify_artifact_with_sources,
+    verify_source_digests,
+};
